ForkExec: better debug output
Include the pid of the child process in all messages relating to it,
to help in cases where the same program is started multiple times.

Log something in the parent when destroying the fake API instance and
thus canceling the fake method call from the child, because that will
trigger a "lost connection" message in the child which was hard to
correlate with events in the server without those additional log
messages in the parent.