SyncContext: include process name in Synthesis log
The process name has become essential to distinguish messages
from local sync parent and child, because the parent will include
the child's output in its own log.

We cannot pass a process name to the Synthesis logging mechanism, so
use the logging prefix instead. Relies on a patch to libsynthesis to
enable the logging of that prefix.