PIM testing: fix testView and testViewSorting
The assertEqual() expected 'quiesence' to be seen, but runUntil() was not
asked to run until that was seen. Therefore tests failed randomly depending on
timing: they passed if the quiescence signal was processed in the same event
loop run as the change that was waited for and failed if it arrived later.

Fixed by waiting for 'quiesence' explicitly. It should arrive last.