testing: fix Exchange eds_contact:testItems
Since the "EDS contacts: avoid unnecessary DB writes during slow sync due to
FILE-AS" change, eds_contact::testItems failed because the "bday;special UTC
offset" (specific to Exchange testing) test user had no FILE-AS set, while the
one eventually stored in EDS had (because of the slow sync patch). Using EDS
3.8 probably would have triggered the same issue.

Fixing this by adding the FILE-AS property to the reference data. This is
a hard requirement now for tests to pass with EDS.